requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.10.6
httpx==0.24.1
python-dotenv==1.0.0
pillow==10.0.0
openai==1.3.0 

boto3==1.34.152
botocore==1.34.152
pynamodb==6.0.1
PyJWT[crypto]==2.8.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
email-validator>=2.0.0
pydantic_core==2.27.2
pynamodb==6.0.1
